Set type of requested_frames of all mixer callbacks to int
The function pointer expects an int so using anything else is either misleading and an implicit cast at best or possibly undefined behavior at worst.
In edge cases, the callbacks using uint16_t can also overflow so this change prevents that bug as well.

+++ src/hardware/innovation.h
@@ -47,7 +47,7 @@ public:
 
 private:
 	bool MaybeRenderFrame(float &frame);
-	void AudioCallback(const uint16_t requested_frames);
+	void AudioCallback(const int requested_frames);
 	uint8_t ReadFromPort(io_port_t port, io_width_t width);
 	void RenderUpToNow();
 	int16_t TallySilence(const int16_t sample);
